tests/wpt/tests/shadow-dom/untriaged/elements-and-dom-objects/extensions-to-element-interface/attributes/test-006.html

<!DOCTYPE html>
<html>
<head>
<title>Shadow DOM Test: A_10_02_01_06</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#shadow-aware-attributes">
<meta name="assert" content="Extensions to Element Interface: shadowRoot of type ShadowRoot">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../../html/resources/common.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function () {

    var d = newHTMLDocument();

    var host = d.createElement('div');
    d.body.appendChild(host);

    assert_equals(host.shadowRoot, null, 'attribute shadowRoot must return null if no shadow tree is accesible');


}, 'A_10_02_01_06_T01');
</script>
</body>
</html>
